+++ odoo/local-src/specific_base/models/__init__.py
@@ -8,3 +8,7 @@ from . import sale_order
 from . import calendar_event
 from . import pricelist
 from . import account_invoice
+from . import res_users
+from . import crm_lead
+from . import utm
+from . import res_country