+++ HISTORY.rst
@@ -8,6 +8,7 @@ Unreleased
 
 **Features and Improvements**
 
+* Card 193: Add specific fields on several models
 * Card 223: Inverse phototherapist and reference fields position on account invoice form view
 * Card 237: Add several modules